requirements.txt
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
scikit-learn==1.4.0
pandas==2.2.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
opencv-python==4.9.0.80
matplotlib==3.8.2
seaborn==0.13.2
tqdm==4.66.1
imbalanced-learn==0.11.0
gunicorn==21.2.0
psutil==5.9.8
h5py==3.10.0
protobuf==4.25.2
tensorflow-cpu==2.15.0
keras==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-hub==0.15.0
tensorflow-addons==0.21.0
python-json-logger==2.0.7
prometheus-client==0.19.0
sentry-sdk==1.39.1
python-dotenv==1.0.1
cachetools==5.3.2
tenacity==8.2.3 